app/requirements.txt
Django==4.2
django-split-settings==1.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
